Drop old contact API.
Remove contacts from global and release component end-points.
Remove role contacts API.
Delete tests for those end-points.
Remove deprecation notices in modified resources.
Move all contact related viewsets, serializers, filters and tests to pdc.app.contact application.
Document and test bulk update on release components

JIRA: PDC-1050
