+++ pdc/apps/component/views.py
@@ -19,10 +19,6 @@ from rest_framework.response import Response
 
 from pdc.apps.common import viewsets
 from pdc.apps.common.models import Label
-from pdc.apps.contact.models import (RoleContact,
-                                     ContactRole,
-                                     GlobalComponentContact,
-                                     ReleaseComponentContact)
 from pdc.apps.common.serializers import LabelSerializer, StrictSerializerMixin
 from pdc.apps.common.filters import LabelFilter
 from .models import (GlobalComponent,
@@ -35,46 +31,19 @@ from .models import (GlobalComponent,
                      ReleaseComponentRelationshipType)
 from .serializers import (GlobalComponentSerializer,
                           ReleaseComponentSerializer,
-                          HackedContactSerializer,
                           BugzillaComponentSerializer,
                           GroupSerializer,
                           GroupTypeSerializer,
                           ReleaseComponentRelationshipSerializer,
                           ReleaseComponentTypeSerializer,
-                          RCRelationshipTypeSerializer,
-                          GlobalComponentContactSerializer,
-                          ReleaseComponentContactSerializer)
+                          RCRelationshipTypeSerializer)
 from .filters import (ComponentFilter,
                       ReleaseComponentFilter,
-                      RoleContactFilter,
                       BugzillaComponentFilter,
                       GroupFilter,
                       GroupTypeFilter,
-                      ReleaseComponentRelationshipFilter,
-                      GlobalComponentContactFilter,
-                      ReleaseComponentContactFilter)
+                      ReleaseComponentRelationshipFilter)
 from . import signals
-
-
-class HackedComponentContactMixin(object):
-
-    def perform_create(self, serializer):
-        # NOTE(xchu): The Creation of ComponentContact is a UPDATE to
-        #             Component instance, the changes about contact have
-        #             been recorded by `serializer.to_internal_value`, so we
-        #             need to record the change about Component itself instead of
-        #             the creation of serializer's object as defined in the
-        #             `ChangeSetCreateModelMixin`.
-        pk = self.kwargs.get("instance_pk", None)
-        component = get_object_or_404(self.model, id=pk)
-        old_value = json.dumps(component.export())
-
-        serializer.save()
-
-        request = self.get_serializer_context().get('request', None)
-        if request and request.changeset:
-            request.changeset.add(self.model.__name__.lower(), component.pk,
-                                  old_value, json.dumps(component.export()))
 
 
 class GlobalComponentViewSet(viewsets.PDCModelViewSet):
@@ -106,13 +75,8 @@ class GlobalComponentViewSet(viewsets.PDCModelViewSet):
     serializer_class = GlobalComponentSerializer
     filter_class = ComponentFilter
 
-    # TODO remove deprecation notices from docstrings once 0.2.0 gets out
     def list(self, request, *args, **kwargs):
         """
-        **Deprecation notice**: the contact field and filters related to it are
-        deprecated and will be remove in next release. Please use [new global
-        component contacts API]($URL:globalcomponentcontacts-list$) instead.
-
         __Method__:
         GET
 
@@ -121,8 +85,6 @@ class GlobalComponentViewSet(viewsets.PDCModelViewSet):
         __Query Params__:
 
         %(FILTERS)s
-
-        The `contact_role` filter must be used together with `email`.
 
         __Response__:
 
@@ -137,17 +99,6 @@ class GlobalComponentViewSet(viewsets.PDCModelViewSet):
                         "name": string,
                         "dist_git_path": <string|null>,
                         "dist_git_web_url": string,
-                        "contacts": [
-                            {
-                                "url": string,
-                                "contact_role": string,
-                                "contact": {
-                                    "mail_name": string,
-                                    "email": string
-                                }
-                            },
-                            ......
-                        ],
                         "labels": [
                             string, ......
                         ],
@@ -164,10 +115,6 @@ class GlobalComponentViewSet(viewsets.PDCModelViewSet):
 
     def retrieve(self, request, *args, **kwargs):
         """
-        **Deprecation notice**: the contact field and filters related to it are
-        deprecated and will be remove in next release. Please use [new global
-        component contacts API]($URL:globalcomponentcontacts-list$) instead.
-
         __Method__:
         GET
 
@@ -180,17 +127,6 @@ class GlobalComponentViewSet(viewsets.PDCModelViewSet):
                 "name": string,
                 "dist_git_path": <string|null>,
                 "dist_git_web_url": string,
-                "contacts": [
-                    {
-                        "url": string,
-                        "contact_role": string,
-                        "contact": {
-                            "mail_name": string,
-                            "email": string
-                        }
-                    },
-                    ......
-                ],
                 "labels": [
                     string, ......
                 ],
@@ -205,10 +141,6 @@ class GlobalComponentViewSet(viewsets.PDCModelViewSet):
 
     def create(self, request, *args, **kwargs):
         """
-        **Deprecation notice**: the contact field and filters related to it are
-        deprecated and will be remove in next release. Please use [new global
-        component contacts API]($URL:globalcomponentcontacts-list$) instead.
-
         __Method__:
         POST
 
@@ -229,7 +161,6 @@ class GlobalComponentViewSet(viewsets.PDCModelViewSet):
                 "name": string,
                 "dist_git_path": string,
                 "dist_git_web_url": string,
-                "contacts": array,
                 "labels": array,
                 "upstream": dict
             }
@@ -244,7 +175,6 @@ class GlobalComponentViewSet(viewsets.PDCModelViewSet):
                 "name": "Demo",
                 "dist_git_path": "rpm/Demo",
                 "dist_git_web_url": "http://pkgs.example.com/cgit/rpms/rpm/Demo"
-                "contacts": [],
                 "labels": [],
                 "upstream": null | {
                                 "homepage": string,
@@ -257,10 +187,6 @@ class GlobalComponentViewSet(viewsets.PDCModelViewSet):
 
     def update(self, request, *args, **kwargs):
         """
-        **Deprecation notice**: the contact field and filters related to it are
-        deprecated and will be remove in next release. Please use [new global
-        component contacts API]($URL:globalcomponentcontacts-list$) instead.
-
         __Method__:
 
         PUT/PATCH: for update
@@ -284,7 +210,6 @@ class GlobalComponentViewSet(viewsets.PDCModelViewSet):
                 "name": string,
                 "dist_git_path": string,
                 "dist_git_web_url": string,
-                "contacts": array,
                 "labels": array,
                 "upstream": null | {
                                 "homepage": string,
@@ -305,7 +230,6 @@ class GlobalComponentViewSet(viewsets.PDCModelViewSet):
                 "name": "Demo",
                 "dist_git_path": "rpm/Demo1",
                 "dist_git_web_url": "http://pkgs.example.com/cgit/rpms/rpm/Demo1"
-                "contacts": [],
                 "labels": [],
                 "upstream": null
             }
@@ -319,7 +243,6 @@ class GlobalComponentViewSet(viewsets.PDCModelViewSet):
                 "name": "Demo",
                 "dist_git_path": "rpm/Demo1",
                 "dist_git_web_url": "http://pkgs.example.com/cgit/rpms/rpm/Demo1"
-                "contacts": [],
                 "labels": [],
                 "upstream": null
             }
@@ -342,160 +265,6 @@ class GlobalComponentViewSet(viewsets.PDCModelViewSet):
             curl -X DELETE -H "Content-Type: application/json" $URL:globalcomponent-detail:4181$
         """
         return super(GlobalComponentViewSet, self).destroy(request, *args, **kwargs)
-
-
-# TODO Remove once 0.2.0 gets out
-class GlobalComponentContactViewSet(HackedComponentContactMixin,
-                                    viewsets.PDCModelViewSet):
-    """
-    This API is deprecated and currently provided only for backwards
-    compatibility. Please use [new global component contacts
-    API]($URL:globalcomponentcontacts-list$).
-    """
-
-    model = GlobalComponent
-    queryset = GlobalComponent.objects.all().order_by('id')
-    serializer_class = HackedContactSerializer
-    filter_class = RoleContactFilter
-
-    def get_queryset(self):
-        gc_id = self.kwargs.get('instance_pk')
-        try:
-            gc = get_object_or_404(GlobalComponent, id=gc_id)
-        except ValueError:
-            # Raised when non-numeric instance_pk is given.
-            raise Http404('Global component with id=%s not found' % gc_id)
-        return gc.contacts.all()
-
-    def list(self, request, *args, **kwargs):
-        """
-        __Method__:
-        GET
-
-        __URL__: $LINK:globalcomponentcontact-list:gc_instance_pk$
-
-        __Query Params__:
-
-        %(FILTERS)s
-
-        The `contact_role` filter must be used together with `email`.
-
-        __Response__:
-
-            [
-                {
-                    "url": url,
-                    "contact_role": string,
-                    "contact": {
-                        <"mail_name"|"username">: string,
-                        "email": string
-                    }
-                },
-                ......
-            ]
-        """
-        # NOTE: we disable pagination because release component contact could
-        #       not do the things like this, and we must keep the outputs same.
-        #       it is a very very dirty hack for current version. we will use
-        #       database view to refactor release component in next sprint.
-        #       then we could enable pagination.
-        self.object_list = self.filter_queryset(self.get_queryset())
-        serializer = self.get_serializer(self.object_list, many=True)
-        return Response(serializer.data)
-
-    def retrieve(self, request, *args, **kwargs):
-        """
-        __Method__:
-        GET
-
-        __URL__: $LINK:globalcomponentcontact-detail:gc_instance_pk:relation_pk$
-
-        __Response__:
-
-            {
-                "url": url,
-                "contact_role": string,
-                "contact": {
-                    <"mail_name"|"username">: string,
-                    "email": string
-                }
-            }
-        """
-        return super(GlobalComponentContactViewSet, self).retrieve(request, *args, **kwargs)
-
-    def update(self, request, *args, **kwargs):
-        return self.http_method_not_allowed(request, *args, **kwargs)
-
-    def bulk_update(self, request, *args, **kwargs):
-        return self.http_method_not_allowed(request, *args, **kwargs)
-
-    def destroy(self, request, *args, **kwargs):
-        """
-        __Method__:
-        DELETE
-
-        __URL__: $LINK:globalcomponentcontact-detail:gc_instance_pk:relation_pk$
-
-        __Response__:
-
-            STATUS: 204 NO CONTENT
-
-        __Example__:
-
-            curl -X DELETE -H "Content-Type: application/json" $URL:globalcomponentcontact-detail:1:1$
-        """
-        pk = self.kwargs.get('pk')
-        contact = get_object_or_404(RoleContact, id=pk)
-
-        gc_id = self.kwargs.get('instance_pk')
-        gc = get_object_or_404(GlobalComponent, id=gc_id)
-
-        old_value = json.dumps(gc.export())
-        if contact in gc.contacts.all():
-            gc.contacts.remove(contact)
-            request.changeset.add("globalcomponent", gc.pk, old_value, json.dumps(gc.export()))
-            return Response(status=status.HTTP_204_NO_CONTENT)
-        else:
-            return Response(
-                data={'detail': 'Can not be deleted since Contact[%s] is not in GlobalComponent[%s].' % (pk, gc_id)},
-                status=status.HTTP_404_NOT_FOUND)
-
-    def create(self, request, *args, **kwargs):
-        """
-        __Method__:
-        POST
-
-        __URL__: $LINK:globalcomponentcontact-list:gc_instance_pk$
-
-        __Data__:
-
-            {
-                'contact_role':                       string,         # required
-                'contact': {
-                    <'username'|'mail_name'>:         string,         # required
-                    'email:                           string,         # required
-                }
-            }
-        __Response__:
-
-            {
-                "url": url,
-                "contact_role": string,
-                "contact": {
-                    <"mail_name"|"username">: string,
-                    "email": string
-                }
-            }
-        """
-        return super(GlobalComponentContactViewSet, self).create(request, *args, **kwargs)
-
-    def get_serializer_context(self):
-        super_class = super(GlobalComponentContactViewSet, self)
-        ret = super_class.get_serializer_context()
-        ret.update({
-            'extra_kwargs': self.kwargs
-        })
-        return ret
 
 
 class GlobalComponentLabelViewSet(viewsets.PDCModelViewSet):
@@ -688,13 +457,8 @@ class ReleaseComponentViewSet(viewsets.PDCModelViewSet):
             qs = qs.filter(release__active=True)
         return qs
 
-    # TODO remove deprecation notices from docstrings once 0.2.0 gets out
     def list(self, request, *args, **kwargs):
         """
-        **Deprecation notice**: the contact field and filters related to it are
-        deprecated and will be remove in next release. Please use [new release
-        component contacts API]($URL:releasecomponentcontacts-list$) instead.
-
         __Method__:
         GET
 
@@ -706,8 +470,6 @@ class ReleaseComponentViewSet(viewsets.PDCModelViewSet):
 
         Components for inactive releases are not shown by default. You can
         change this with `include_inactive_release` set to any non-empty value.
-
-        The `contact_role` filter must be used together with `email`.
 
         __Response__:
 
@@ -737,9 +499,6 @@ class ReleaseComponentViewSet(viewsets.PDCModelViewSet):
                         "dist_git_branch": <string|null>,
                         "dist_git_web_url": string,
                         "active": bool,
-                        "contacts": [
-                            ......
-                        ],
                         "srpm": {
                             "name": string
                         } OR null,
@@ -752,10 +511,6 @@ class ReleaseComponentViewSet(viewsets.PDCModelViewSet):
 
     def retrieve(self, request, *args, **kwargs):
         """
-        **Deprecation notice**: the contact field and filters related to it are
-        deprecated and will be remove in next release. Please use [new release
-        component contacts API]($URL:releasecomponentcontacts-list$) instead.
-
         __Method__:
         GET
 
@@ -783,29 +538,18 @@ class ReleaseComponentViewSet(viewsets.PDCModelViewSet):
                 "dist_git_branch": <string|null>,
                 "dist_git_web_url": string,
                 "active": true,
-                "contacts": [
-                    ......
-                ],
                 "srpm": {
                     "name": string
                 } OR null,
                 "type": <string|null>
             }
 
-        __NOTE__:
-
-        ``contacts`` field will inherit from corresponding global component
-        contacts(with ``"inherited": true``) if any.
         """
 
         return super(ReleaseComponentViewSet, self).retrieve(request, *args, **kwargs)
 
     def update(self, request, *args, **kwargs):
         """
-        **Deprecation notice**: the contact field and filters related to it are
-        deprecated and will be remove in next release. Please use [new release
-        component contacts API]($URL:releasecomponentcontacts-list$) instead.
-
         __Method__:
 
         PUT:
@@ -869,9 +613,6 @@ class ReleaseComponentViewSet(viewsets.PDCModelViewSet):
                 "dist_git_branch": <string|null>,
                 "dist_git_web_url": string,
                 "active": true,
-                "contacts": [
-                    ......
-                ],
                 "srpm": {
                     "name": string
                 } OR null
@@ -898,10 +639,6 @@ class ReleaseComponentViewSet(viewsets.PDCModelViewSet):
 
     def create(self, request, *args, **kwargs):
         """
-        **Deprecation notice**: the contact field and filters related to it are
-        deprecated and will be remove in next release. Please use [new release
-        component contacts API]($URL:releasecomponentcontacts-list$) instead.
-
         __Method__:
         POST
 
@@ -948,9 +685,6 @@ class ReleaseComponentViewSet(viewsets.PDCModelViewSet):
                         ......
                     ]
                 },
-                "contacts": [
-                    ......
-                ],
                 "srpm": {
                     "name": string
                 } OR null,
@@ -988,354 +722,12 @@ class ReleaseComponentViewSet(viewsets.PDCModelViewSet):
         """
         return super(ReleaseComponentViewSet, self).destroy(request, *args, **kwargs)
 
-    # TODO remove this method once 0.2.0 gets out, regular bulk update should be used
-    def bulk_update(self, request, *args, **kwargs):
-        """
-        **Deprecation notice**: this method is deprecated and will be removed
-        in next release. Please use [new release component contacts
-        API]($URL:releasecomponentcontacts-list$) instead.
-
-        __Method__:
-        PUT
-
-        __URL__: $LINK:releasecomponent-list$
-
-        __Data__:
-
-            {
-                'releases':                                  array,          # optional
-                'global_component':                          string,         # required
-                'contacts': [
-                    {
-                        'contact_role':                      string,         # required
-                        'contact': {
-                            <'username'|'mail_name'>:        string,         # required
-                            'email:                          string,         # required
-                        }
-                    },
-                    ......
-                ]
-            }
-        __Response__:
-
-            [
-                {
-                    "id": int,
-                    "release": {
-                        "release_id": string,
-                        "active": bool
-                    },
-                    "global_component": string,
-                    "bugzilla_component": string,
-                    "brew_package": string
-                    "name": string,
-                    "dist_git_branch": <string|null>,
-                    "dist_git_web_url": string,
-                    "active": true,
-                    "contacts": [
-                        ......
-                    ],
-                    "srpm": {
-                        "name": string,
-                    } OR null,
-                    "type": <string|null>
-                },
-                ......
-            ]
-        """
-        required_fields = set(("global_component", "contacts"))
-        missing_fields = required_fields - set(request.data.keys())
-        if missing_fields:
-            return Response(status=status.HTTP_400_BAD_REQUEST,
-                            data={'detail': 'Missing arguments: %s' % ', '.join(missing_fields)})
-
-        global_component_name = request.data.get('global_component')
-        releases = request.data.get('releases')
-        contacts = request.data.get('contacts')
-
-        if not global_component_name:
-            return Response({"detail": "global_component is required."}, status=status.HTTP_400_BAD_REQUEST)
-
-        qs = self.get_queryset().filter(global_component__name=global_component_name)
-
-        if releases:
-            release = (releases,) if not isinstance(releases, list) else releases
-            qs = qs.filter(release__release_id__in=release)
-
-        # iterate release component records
-        for row in qs.iterator():
-            context = self.get_serializer_context()
-            context.update({'extra_kwargs': {'instance_pk': row.id}})
-            contact_serializer = HackedContactSerializer(data=contacts,
-                                                         many=True,
-                                                         context=context)
-
-            if not contact_serializer.is_valid():
-                return Response(contact_serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-
-            # contact does not exist in this release component
-            # dump release component before update
-            pre_export = row.export()
-            # iterate new contacts.
-            for obj in contact_serializer.validated_data:
-                if not row.contacts.filter(pk=obj.pk).exists():
-                    overwrite_contact = row.contacts.filter(contact_role_id=obj.contact_role_id)
-                    if overwrite_contact.exists():
-                        # release component has the same contact type contact.
-                        # drop it and insert new one.
-                        overwrite_contact.delete()
-                    # save new contact.
-                    row.contacts.add(obj)
-            request.changeset.add("releasecomponent",
-                                  row.pk,
-                                  json.dumps(pre_export),
-                                  json.dumps(row.export()))
-        serializer = ReleaseComponentSerializer(instance=qs, many=True, context={'request': request})
-        return Response(serializer.data, status=status.HTTP_200_OK)
-
-    # TODO enable once 0.2.0 gets out
     def bulk_partial_update(self, request):
         # Without this definition, bulk_operations would pick up PATCH
         # requests, convert them to PUT with kwargs['partial'] = True and an
         # error about potentially missing keys would be returned. This way it
         # says PATCH is not allowed.
         return self.http_method_not_allowed(request)
-
-
-# TODO Remove once 0.2.0 gets out
-class ReleaseComponentContactViewSet(HackedComponentContactMixin,
-                                     viewsets.PDCModelViewSet):
-    """
-    This API is deprecated and currently provided only for backwards
-    compatibility. Please use [new release component contacts
-    API]($URL:releasecomponentcontacts-list$) instead.
-    """
-    model = ReleaseComponent
-    queryset = ReleaseComponent.objects.all().order_by('id')
-    serializer_class = HackedContactSerializer
-    gcc_serializer_class = HackedContactSerializer
-    extra_query_params = ('contact_role', )
-
-    def get_queryset(self):
-        rc_id = self.kwargs.get("instance_pk", None)
-        try:
-            release_component = get_object_or_404(ReleaseComponent, pk=rc_id)
-        except ValueError:
-            # Raised when non-numeric instance_pk is given.
-            raise Http404('Release component with id=%s not found' % rc_id)
-        return release_component.contacts.all()
-
-    def get_serializer_context(self):
-        super_class = super(ReleaseComponentContactViewSet, self)
-        ret = super_class.get_serializer_context()
-        ret.update({
-            'extra_kwargs': self.kwargs
-        })
-        return ret
-
-    def get_serializer(self, *args, **kwargs):
-        serializer_class = self.get_serializer_class()
-        kwargs['context'] = self.get_serializer_context()
-        kwargs['view_name'] = 'releasecomponentcontact-detail'
-        return serializer_class(*args, **kwargs)
-
-    def list(self, request, *args, **kwargs):
-        """
-        __Method__:
-        GET
-
-        __URL__: $LINK:releasecomponentcontact-list:rc_instance_pk$
-
-        __Query Params__:
-
-        %(FILTERS)s
-
-        __Response__:
-
-            [
-                {
-                    "url": url,
-                    "contact_role": string,
-                    "contact": {
-                        <"mail_name"|"username">: string,
-                        "email": string
-                    }
-                },
-                ......
-            ]
-        """
-        # FIXME: Design issue here. Contact inheritance(PDC-184) requires access
-        # of different resource under same url. e.g., could view global component
-        # contact information under release component contact list.
-        # This is not RESTful. To meet this needs, I have to hard code this
-        # `ViewSet.list`, as I failed to find an elegant way. Unfortunately we
-        # cannot benefit from DRF built-in features such as pagination.
-        rc_id = kwargs.get("instance_pk", None)
-        # Release Component object
-        try:
-            rc = get_object_or_404(ReleaseComponent, pk=rc_id)
-        except ValueError:
-            # Raised when non-numeric instance_pk is given.
-            raise Http404('Release component with id=%s not found' % rc_id)
-        gcc_qs = GlobalComponent.objects.get(pk=rc.global_component_id).contacts.all()
-        rcc_qs = rc.contacts.all()
-        # Contact type based inheritance mechanisms(excerpted from JIRA PDC-184)
-        # - Input without contact_role, output contacts including
-        #   release_component's contacts and global_component's contacts which
-        #   is not in self contacts.
-        # - Input with specific contact_role which is only in global_component
-        #   output contact will inherit from global_component's.
-        # - Input with specific contact_role which is in release_component,
-        #   output contact will be it's own contact
-        rcc_types = map(lambda s: s.contact_role.name, rcc_qs)
-        # excluded_types: existing contact types for release component
-        excluded_types = ContactRole.objects.filter(name__in=rcc_types)
-        # exclude co-exists(same contact type) contacts from global component.
-        gcc_qs = gcc_qs.exclude(contact_role__in=excluded_types)
-
-        query_params = request.query_params
-        contact_role = query_params.getlist('contact_role')
-        # FIXME: add more options in query parameter.
-        # username = query_params.getlist('username')
-        # mail_name = query_params.getlist('mail_name')
-        # email = query_params.getlist('email')
-        if contact_role:
-            c_types = ContactRole.objects.filter(name__in=contact_role)
-            gcc_qs = gcc_qs.filter(contact_role__in=c_types)
-            rcc_qs = rcc_qs.filter(contact_role__in=c_types)
-
-        context = self.get_serializer_context()
-        view_name = 'releasecomponentcontact-detail'
-        serializer = self.serializer_class(rcc_qs, many=True,
-                                           context=context,
-                                           view_name=view_name)
-        ret = serializer.data
-        # FIXME: We need to add identity `inherited=True` for global component
-        # contacts in the resulting release component contacts(PDC-166).
-        # Since the attribute `inherited` only makes sense for release component
-        # contacts, it's not appropriate to hack this `inherited=True` stuff from
-        # within the GlobalComponentContactSerializer.
-        # Below is an ugly hack over GlobalComponentContactSerializer.data.
-        # Note that with this change, the resulting contact representations will
-        # be in-consistent accross global/release component contacts.
-        # We need to access inconsistent resources under a specific URL,
-        # in our case, we need to access release component contacts as well as
-        # global component contact (if any) under the same url.
-        # This is non-semantic URL design [1] and needs a better approach.
-        # [1] http://en.wikipedia.org/wiki/Semantic_URL
-        gcc_serializer = self.gcc_serializer_class(
-            gcc_qs,
-            many=True,
-            inherited=True,
-            # NOTE(xchu): The `HackedContactSerializer` used here need `instance_pk`
-            #             in the context `extra_kwargs` to build the 'url' in response.
-            #             For the inherited contacts, we need to pass
-            #             `global_component_id` as `instance_pk`.
-            context={'request': request,
-                     'extra_kwargs': {'instance_pk': rc.global_component_id}})
-        gc_contacts = gcc_serializer.data
-        ret.extend(gc_contacts)
-
-        return Response(ret)
-
-    def retrieve(self, request, *args, **kwargs):
-        """
-        __Method__:
-        GET
-
-        __URL__: $LINK:releasecomponentcontact-detail:rc_instance_pk:relation_pk$
-
-        __Response__:
-
-            {
-                "url": url,
-                "contact_role": string,
-                "contact": {
-                    <"mail_name"|"username">: string,
-                    "email": string
-                }
-            }
-        """
-        return super(ReleaseComponentContactViewSet, self).retrieve(request, *args, **kwargs)
-
-    def create(self, request, *args, **kwargs):
-        """
-        __Method__:
-        POST
-
-        __URL__: $LINK:releasecomponentcontact-list:rc_instance_pk$
-
-        __Data__:
-
-            {
-                'contact_role':                       string,         # required
-                'contact': {
-                    <'username'|'mail_name'>:         string,         # required
-                    'email:                           string,         # required
-                }
-            }
-        __Response__:
-
-            {
-                "url": url,
-                "contact_role": string,
-                "contact": {
-                    <"mail_name"|"username">: string,
-                    "email": string
-                }
-            }
-        """
-        return super(ReleaseComponentContactViewSet, self).create(request, *args, **kwargs)
-
-    def destroy(self, request, *args, **kwargs):
-        """
-        __Method__:
-        DELETE
-
-        __URL__: $LINK:releasecomponentcontact-detail:rc_instance_pk:relation_pk$
-
-        __Response__:
-
-            STATUS: 204 NO CONTENT
-
-        __Example__:
-
-            curl -X DELETE -H "Content-Type: application/json" $URL:releasecomponentcontact-detail:2:3175$
-        """
-        pk = self.kwargs.get('pk')
-        try:
-            contact = RoleContact.objects.get(id=pk)
-        except RoleContact.DoesNotExist:
-            return Response(data={'detail': 'Specified TypedContact[%s] not found.' % pk},
-                            status=status.HTTP_404_NOT_FOUND)
-
-        rc_id = self.kwargs.get('instance_pk')
-        try:
-            rc = ReleaseComponent.objects.get(id=rc_id)
-        except ReleaseComponent.DoesNotExist:
-            return Response(data={'detail': 'Specified ReleaseComponent[%s] not found.' % rc_id},
-                            status=status.HTTP_404_NOT_FOUND)
-
-        old_value = json.dumps(rc.export())
-        if contact in rc.contacts.all():
-            rc.contacts.remove(contact)
-            request.changeset.add("releasecomponent", rc.pk, old_value, json.dumps(rc.export()))
-            return Response(status=status.HTTP_204_NO_CONTENT)
-        elif contact in rc.global_component.contacts.all():
-            return Response(
-                data={'detail': 'Contact[%s] is inherited from GlobalComponent[%s] and can only be deleted from there.'
-                      % (pk, rc.global_component.pk)},
-                status=status.HTTP_400_BAD_REQUEST)
-        else:
-            return Response(
-                data={'detail': 'Can not be deleted since Contact[%s] is not in ReleaseComponent[%s].' % (pk, rc_id)},
-                status=status.HTTP_404_NOT_FOUND)
-
-    def update(self, request, *args, **kwargs):
-        return self.http_method_not_allowed(request, *args, **kwargs)
-
-    def bulk_update(self, request, *args, **kwargs):
-        return self.http_method_not_allowed(request, *args, **kwargs)
 
 
 class BugzillaComponentViewSet(viewsets.PDCModelViewSet):
@@ -1899,127 +1291,3 @@ class ReleaseComponentRelationshipViewSet(viewsets.PDCModelViewSet):
         On success, HTTP status code is 204 and the response has no content.
         """
         return super(ReleaseComponentRelationshipViewSet, self).destroy(request, *args, **kwargs)
-
-
-class _BaseContactViewSet(viewsets.PDCModelViewSet):
-    def list(self, *args, **kwargs):
-        """
-        __Method__: `GET`
-
-        __URL__: $LINK:%(BASENAME)s-list$
-
-        __Query params__:
-
-        %(FILTERS)s
-
-        The value of `contact` filter should either be a username or mailling
-        list name.
-
-        __Response__: a paged list of following objects
-
-        %(SERIALIZER)s
-        """
-        return super(_BaseContactViewSet, self).list(*args, **kwargs)
-
-    def retrieve(self, *args, **kwargs):
-        """
-        __Method__: `GET`
-
-        __URL__: $LINK:%(BASENAME)s-detail:pk$
-
-        __Response__:
-
-        %(SERIALIZER)s
-        """
-        return super(_BaseContactViewSet, self).retrieve(*args, **kwargs)
-
-    def destroy(self, *args, **kwargs):
-        """Remove association between component and contact.
-
-        __Method__: `DELETE`
-
-        __URL__: $LINK:%(BASENAME)s-detail:pk$
-
-        __Response__: Nothing on success.
-        """
-        return super(_BaseContactViewSet, self).destroy(*args, **kwargs)
-
-    def update(self, *args, **kwargs):
-        """Change details about a contact linked to component.
-
-        Please not that if you change the `contact` field here, only the single
-        updated relationship between contact and component will be updated.
-        Specifically, no other component will be affected.
-
-        If you update with new contact details and such contact does not exist
-        yet, it will be automatically created. The specific type will be chosen
-        based on whether `username` or `mail_name` was used.
-
-        __Method__: `PUT`, `PATCH`
-
-        __URL__: $LINK:%(BASENAME)s-detail:pk$
-
-        __Data__:
-
-        %(WRITABLE_SERIALIZER)s
-
-        %(WRITABLE_DATA_COMMENT)s
-
-        View [list of available contact roles]($URL:contactrole-list$).
-
-        __Response__:
-
-        %(SERIALIZER)s
-        """
-        return super(_BaseContactViewSet, self).update(*args, **kwargs)
-
-    def create(self, *args, **kwargs):
-        """Connect contact details with a component.
-
-        If the contact does not exist, it will be created automatically.
-
-        __Method__: `POST`
-
-        __URL__: $LINK:%(BASENAME)s-list$
-
-        __Data__:
-
-        %(WRITABLE_SERIALIZER)s
-
-        %(WRITABLE_DATA_COMMENT)s
-
-        Depending on whether `username` or `mail_name` is used, a person or
-        mailling list will be linked to the component.
-
-        View [list of available contact roles]($URL:contactrole-list$).
-
-        __Response__:
-
-        %(SERIALIZER)s
-        """
-        return super(_BaseContactViewSet, self).create(*args, **kwargs)
-
-
-# TODO Remove Info from name once 0.2.0 gets out
-class GlobalComponentContactInfoViewSet(_BaseContactViewSet):
-
-    queryset = GlobalComponentContact.objects.all().select_related().order_by('id')
-    serializer_class = GlobalComponentContactSerializer
-    filter_class = GlobalComponentContactFilter
-    docstring_macros = {
-        'BASENAME': 'globalcomponentcontacts',
-        'WRITABLE_DATA_COMMENT': '',
-    }
-
-
-# TODO Remove Info from name once 0.2.0 gets out
-class ReleaseComponentContactInfoViewSet(_BaseContactViewSet):
-
-    queryset = ReleaseComponentContact.objects.all().select_related().order_by('id')
-    serializer_class = ReleaseComponentContactSerializer
-    filter_class = ReleaseComponentContactFilter
-    docstring_macros = {
-        'BASENAME': 'releasecomponentcontacts',
-        'WRITABLE_DATA_COMMENT': 'The component can be alternatively specified ' +
-                                 'by its id as `{"id": "int"}`.',
-    }
