+++ pdc/apps/contact/tests.py
@@ -3,20 +3,22 @@
 # Licensed under The MIT License (MIT)
 # http://opensource.org/licenses/MIT
 #
-import mock
 
 from django.core.urlresolvers import reverse
-from django.db.models import ProtectedError
 
 from rest_framework import status
 from rest_framework.test import APITestCase
 
 from pdc.apps.common.test_utils import TestCaseWithChangeSetMixin
-from .models import ContactRole, RoleContact, Person
+from .models import ContactRole, Person, Maillist, GlobalComponentContact, ReleaseComponentContact
 
 
 class ContactRoleRESTTestCase(TestCaseWithChangeSetMixin, APITestCase):
-    fixtures = ['pdc/apps/contact/fixtures/tests/contact_role.json', ]
+    fixtures = ['pdc/apps/component/fixtures/tests/global_component.json',
+                'pdc/apps/contact/fixtures/tests/contact_role.json',
+                'pdc/apps/contact/fixtures/tests/person.json',
+                'pdc/apps/component/fixtures/tests/upstream.json'
+                ]
 
     def test_create(self):
         url = reverse('contactrole-list')
@@ -86,8 +88,11 @@ class ContactRoleRESTTestCase(TestCaseWithChangeSetMixin, APITestCase):
         self.assertNumChanges([1])
 
     def test_delete_protect(self):
-        RoleContact.specific_objects.create(username='person1', email='person1@test.com',
-                                            contact_role='qe_ack')
+        from pdc.apps.component import models as component_models
+        GlobalComponentContact.objects.create(
+            component=component_models.GlobalComponent.objects.get(name='MySQL-python'),
+            role=ContactRole.objects.get(name='qe_ack'),
+            contact=Person.objects.get(username='person1'))
 
         url = reverse('contactrole-detail', args=['qe_ack'])
         response = self.client.delete(url, format='json')
@@ -121,7 +126,11 @@ class ContactRoleRESTTestCase(TestCaseWithChangeSetMixin, APITestCase):
 
 
 class PersonRESTTestCase(TestCaseWithChangeSetMixin, APITestCase):
-    fixtures = ['pdc/apps/contact/fixtures/tests/person.json', ]
+    fixtures = ['pdc/apps/component/fixtures/tests/global_component.json',
+                'pdc/apps/contact/fixtures/tests/contact_role.json',
+                'pdc/apps/contact/fixtures/tests/person.json',
+                'pdc/apps/component/fixtures/tests/upstream.json'
+                ]
 
     def test_create(self):
         url = reverse('person-list')
@@ -260,8 +269,11 @@ class PersonRESTTestCase(TestCaseWithChangeSetMixin, APITestCase):
         self.assertNumChanges([1])
 
     def test_delete_protect(self):
-        RoleContact.specific_objects.create(username='person1', email='person1@test.com',
-                                            contact_role='qe_ack')
+        from pdc.apps.component import models as component_models
+        GlobalComponentContact.objects.create(
+            component=component_models.GlobalComponent.objects.get(name='MySQL-python'),
+            role=ContactRole.objects.get(name='qe_ack'),
+            contact=Person.objects.get(username='person1'))
 
         url = reverse('person-detail', args=[3])
         response = self.client.delete(url, format='json')
@@ -271,7 +283,11 @@ class PersonRESTTestCase(TestCaseWithChangeSetMixin, APITestCase):
 
 
 class MaillistRESTTestCase(TestCaseWithChangeSetMixin, APITestCase):
-    fixtures = ['pdc/apps/contact/fixtures/tests/maillist.json', ]
+    fixtures = ['pdc/apps/component/fixtures/tests/global_component.json',
+                'pdc/apps/contact/fixtures/tests/contact_role.json',
+                'pdc/apps/contact/fixtures/tests/maillist.json',
+                'pdc/apps/component/fixtures/tests/upstream.json'
+                ]
 
     def test_create(self):
         url = reverse('maillist-list')
@@ -367,8 +383,11 @@ class MaillistRESTTestCase(TestCaseWithChangeSetMixin, APITestCase):
         self.assertNumChanges([1])
 
     def test_delete_protect(self):
-        RoleContact.specific_objects.create(mail_name='maillist1', email='maillist1@test.com',
-                                            contact_role='qe_ack')
+        from pdc.apps.component import models as component_models
+        GlobalComponentContact.objects.create(
+            component=component_models.GlobalComponent.objects.get(name='MySQL-python'),
+            role=ContactRole.objects.get(name='qe_ack'),
+            contact=Maillist.objects.get(mail_name='maillist1'))
 
         url = reverse('maillist-detail', args=[1])
         response = self.client.delete(url, format='json')
@@ -377,354 +396,17 @@ class MaillistRESTTestCase(TestCaseWithChangeSetMixin, APITestCase):
         self.assertNumChanges([])
 
     def test_multi_delete_protect_no_change_set(self):
-        RoleContact.specific_objects.create(mail_name='maillist1', email='maillist1@test.com',
-                                            contact_role='qe_ack')
+        from pdc.apps.component import models as component_models
+        GlobalComponentContact.objects.create(
+            component=component_models.GlobalComponent.objects.get(name='MySQL-python'),
+            role=ContactRole.objects.get(name='qe_ack'),
+            contact=Maillist.objects.get(mail_name='maillist1'))
 
         url = reverse('maillist-detail', args=[1])
         #  try to delete it multi times, verify changes count
         self.client.delete(url, format='json')
         self.client.delete(url, format='json')
         self.client.delete(url, format='json')
-        self.assertNumChanges([])
-
-
-class RoleContactRESTTestCase(TestCaseWithChangeSetMixin, APITestCase):
-
-    def setUp(self):
-        super(RoleContactRESTTestCase, self).setUp()
-
-        RoleContact.specific_objects.create(username='person1', email='person1@test.com', contact_role='qe_ack')
-        RoleContact.specific_objects.create(username='person2', email='person2@test.com', contact_role='pm')
-        RoleContact.specific_objects.create(mail_name='maillist1', email='maillist1@test.com', contact_role='qe_team')
-        RoleContact.specific_objects.create(mail_name='maillist2', email='maillist2@test.com', contact_role='devel_team')
-
-    def test_create_with_new_type_should_not_be_allowed(self):
-        url = reverse('rolecontact-list')
-        data = {'contact': {'username': 'person1', 'email': 'person1@test.com'},
-                'contact_role': 'new_type'}
-        response = self.client.post(url, data, format='json')
-        self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
-
-    def test_create_changeset_with_new_person(self):
-        url = reverse('rolecontact-list')
-        data = {'contact': {'username': 'new_person', 'email': 'new_person@test.com'},
-                'contact_role': 'pm'}
-        response = self.client.post(url, data, format='json')
-
-        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
-        self.assertEqual(response.data.get('contact').get('username'), 'new_person')
-        self.assertEqual(response.data.get('contact').get('email'), 'new_person@test.com')
-        self.assertEqual(response.data.get('contact_role'), 'pm')
-        self.assertNumChanges([2])
-
-    def test_create_with_person(self):
-        url = reverse('rolecontact-list')
-        data = {'contact': {'username': 'test_person', 'email': 'test@test.com'},
-                'contact_role': 'qe_ack'}
-        response = self.client.post(url, data, format='json')
-
-        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
-        self.assertEqual(response.data.get('contact').get('username'), 'test_person')
-        self.assertEqual(response.data.get('contact').get('email'), 'test@test.com')
-        self.assertEqual(response.data.get('contact_role'), 'qe_ack')
-        self.assertNumChanges([2])
-
-    def test_create_with_maillist(self):
-        url = reverse('rolecontact-list')
-        data = {'contact': {'mail_name': 'test_mail', 'email': 'test@test.com'},
-                'contact_role': 'qe_ack'}
-        response = self.client.post(url, data, format='json')
-
-        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
-        self.assertEqual(response.data.get('contact').get('mail_name'), 'test_mail')
-        self.assertEqual(response.data.get('contact').get('email'), 'test@test.com')
-        self.assertEqual(response.data.get('contact_role'), 'qe_ack')
-        self.assertNumChanges([2])
-
-    def test_create_with_wrong_field(self):
-        url = reverse('rolecontact-list')
-        data = {'wrong_name': 'test_type'}
-        response = self.client.post(url, data, format='json')
-
-        self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
-        self.assertEqual(response.data, {'detail': 'Unknown fields: "wrong_name".'})
-        self.assertNumChanges([])
-
-    def test_create_with_missing_field(self):
-        url = reverse('rolecontact-list')
-        response = self.client.post(url, {}, format='json')
-
-        self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
-        self.assertEqual(response.data, {"contact_role": ["This field is required."],
-                                         "contact": ["This field is required."]})
-        self.assertNumChanges([])
-
-    def test_create_with_invalid_object_field(self):
-        url = reverse('rolecontact-list')
-        data = {'contact': {'username': 'person1', 'invalid_key': 'person1@test.com'},
-                'contact_role': 'qe_ack'}
-
-        response = self.client.post(url, data, format='json')
-
-        self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
-        self.assertIn('invalid_key', response.content)
-        self.assertNumChanges([])
-
-    def test_create_with_exists_value(self):
-        url = reverse('rolecontact-list')
-        data = {'contact': {'username': 'person1', 'email': 'person1@test.com'},
-                'contact_role': 'qe_ack'}
-        response = self.client.post(url, data, format='json')
-
-        self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
-        self.assertEqual(response.data, {"detail": ["The fields (\'contact\', \'contact_role\') must make a unique set."]})
-        self.assertNumChanges([])
-
-    def test_create_with_bad_type(self):
-        url = reverse('rolecontact-list')
-        data = {'contact': {'person_name': 'person1', 'e-mail': 'person1@test.com'},
-                'contact_role': 'qe_ack'}
-        response = self.client.post(url, data, format='json')
-
-        self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
-        self.assertEqual(response.data, {"contact": ["Could not determine type of contact."]})
-        self.assertNumChanges([])
-
-    def test_get(self):
-        url = reverse('rolecontact-detail', args=[1])
-        response = self.client.get(url, format='json')
-
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(response.data.get('contact').get('username'), 'person1')
-        self.assertEqual(response.data.get('contact').get('email'), 'person1@test.com')
-        self.assertEqual(response.data.get('contact_role'), 'qe_ack')
-
-    def test_query_with_username(self):
-        url = reverse('rolecontact-list')
-        response = self.client.get(url + '?username=person2', format='json')
-
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(response.data.get('results')[0].get('contact').get('username'), 'person2')
-
-    def test_query_with_mail_name(self):
-        url = reverse('rolecontact-list')
-        response = self.client.get(url + '?mail_name=maillist2', format='json')
-
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(response.data.get('results')[0].get('contact').get('mail_name'), 'maillist2')
-
-    def test_query_with_email(self):
-        url = reverse('rolecontact-list')
-        response = self.client.get(url + '?email=person2@test.com', format='json')
-
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(response.data.get('results')[0].get('contact').get('email'), 'person2@test.com')
-
-    def test_query_with_contact_role(self):
-        url = reverse('rolecontact-list')
-        response = self.client.get(url + '?contact_role=pm', format='json')
-
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(response.data.get('results')[0].get('contact_role'), 'pm')
-
-    def test_query_with_username_list(self):
-        url = reverse('rolecontact-list')
-        response = self.client.get(url + '?username=person1&username=person2', format='json')
-
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(response.data.get('count'), 2)
-        self.assertEqual(response.data.get('results')[0].get('contact').get('username'), 'person1')
-        self.assertEqual(response.data.get('results')[1].get('contact').get('username'), 'person2')
-
-    def test_query_with_mail_name_list(self):
-        url = reverse('rolecontact-list')
-        response = self.client.get(url + '?mail_name=maillist1&mail_name=maillist2', format='json')
-
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(response.data.get('count'), 2)
-        self.assertEqual(response.data.get('results')[0].get('contact').get('mail_name'), 'maillist1')
-        self.assertEqual(response.data.get('results')[1].get('contact').get('mail_name'), 'maillist2')
-
-    def test_query_with_email_list(self):
-        url = reverse('rolecontact-list')
-        response = self.client.get(url + '?email=person1@test.com&email=maillist2@test.com', format='json')
-
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(response.data.get('count'), 2)
-        self.assertEqual(response.data.get('results')[0].get('contact').get('email'), 'person1@test.com')
-        self.assertEqual(response.data.get('results')[1].get('contact').get('email'), 'maillist2@test.com')
-
-    def test_query_with_contact_role_list(self):
-        url = reverse('rolecontact-list')
-        response = self.client.get(url + '?contact_role=qe_ack&contact_role=qe_team', format='json')
-
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(response.data.get('count'), 2)
-        self.assertEqual(response.data.get('results')[0].get('contact_role'), 'qe_ack')
-        self.assertEqual(response.data.get('results')[1].get('contact_role'), 'qe_team')
-
-    def test_query_with_username_mail_name_mixup(self):
-        url = reverse('rolecontact-list')
-        response = self.client.get(url + '?username=person1&mail_name=maillist2', format='json')
-
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(response.data.get('count'), 2)
-        self.assertEqual(response.data.get('results')[0].get('contact').get('username'), 'person1')
-        self.assertEqual(response.data.get('results')[1].get('contact').get('mail_name'), 'maillist2')
-
-    def test_query_with_username_email_mixup(self):
-        url = reverse('rolecontact-list')
-        response = self.client.get(url + '?username=person1&email=maillist2@test.com', format='json')
-
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(response.data.get('count'), 0)
-
-        response = self.client.get(url + '?username=person1&email=person1@test.com', format='json')
-
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(response.data.get('count'), 1)
-        self.assertEqual(response.data.get('results')[0].get('contact').get('username'), 'person1')
-
-    def test_query_with_username_contact_role_mixup(self):
-        url = reverse('rolecontact-list')
-        response = self.client.get(url + '?username=person1&contact_role=pm', format='json')
-
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(response.data.get('count'), 0)
-        response = self.client.get(url + '?username=person1&contact_role=qe_ack', format='json')
-
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(response.data.get('count'), 1)
-        self.assertEqual(response.data.get('results')[0].get('contact').get('username'), 'person1')
-        self.assertEqual(response.data.get('results')[0].get('contact_role'), 'qe_ack')
-
-    def test_query_with_mail_name_email_mixup(self):
-        url = reverse('rolecontact-list')
-        response = self.client.get(url + '?mail_name=maillist1&email=maillist2@test.com', format='json')
-
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(response.data.get('count'), 0)
-
-        response = self.client.get(url + '?mail_name=maillist1&email=maillist1@test.com', format='json')
-
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(response.data.get('count'), 1)
-        self.assertEqual(response.data.get('results')[0].get('contact').get('mail_name'), 'maillist1')
-        self.assertEqual(response.data.get('results')[0].get('contact').get('email'), 'maillist1@test.com')
-
-    def test_query_with_mail_name_contact_role_mixup(self):
-        url = reverse('rolecontact-list')
-        response = self.client.get(url + '?mail_name=maillist1&contact_role=devel_team', format='json')
-
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(response.data.get('count'), 0)
-
-        response = self.client.get(url + '?mail_name=maillist1&contact_role=qe_team', format='json')
-
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(response.data.get('count'), 1)
-        self.assertEqual(response.data.get('results')[0].get('contact').get('mail_name'), 'maillist1')
-        self.assertEqual(response.data.get('results')[0].get('contact_role'), 'qe_team')
-
-    def test_query_with_email_contact_role_mixup(self):
-        url = reverse('rolecontact-list')
-        response = self.client.get(url + '?email=person1@test.com&contact_role=pm', format='json')
-
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(response.data.get('count'), 0)
-
-        response = self.client.get(url + '?email=person1@test.com&contact_role=qe_ack', format='json')
-
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(response.data.get('count'), 1)
-        self.assertEqual(response.data.get('results')[0].get('contact').get('email'), 'person1@test.com')
-        self.assertEqual(response.data.get('results')[0].get('contact_role'), 'qe_ack')
-
-    def test_query_with_multi_key_list(self):
-        url = reverse('rolecontact-list')
-        response = self.client.get(url + '?username=person1&username=person2&contact_role=pm', format='json')
-
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(response.data.get('count'), 1)
-        self.assertEqual(response.data.get('results')[0].get('contact').get('username'), 'person2')
-        response = self.client.get(url + '?username=person1&username=person2&contact_role=pm&contact_role=qe_ack', format='json')
-
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(response.data.get('count'), 2)
-        self.assertEqual(response.data.get('results')[0].get('contact').get('username'), 'person1')
-        self.assertEqual(response.data.get('results')[1].get('contact').get('username'), 'person2')
-
-    def test_patch_update_with_contact_role(self):
-        url = reverse('rolecontact-detail', args=[1])
-        data = {'contact_role': 'pm'}
-        response = self.client.patch(url, data, format='json')
-
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(response.data.get('contact_role'), 'pm')
-        self.assertNumChanges([1])
-
-    def test_patch_update_with_contact(self):
-        url = reverse('rolecontact-detail', args=[1])
-        data = {'contact': {'username': 'new_name', 'email': 'new@test.com'}}
-        response = self.client.patch(url, data, format='json')
-
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(response.data.get('contact').get('username'), 'new_name')
-        self.assertEqual(response.data.get('contact').get('email'), 'new@test.com')
-        self.assertNumChanges([2])
-
-    def test_patch_update_with_bad_contact(self):
-        url = reverse('rolecontact-detail', args=[1])
-        data = {'contact': {'mali_list': 'new_name', 'email': 'new@test.com'}}
-        response = self.client.patch(url, data, format='json')
-
-        self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
-        self.assertEqual(response.data, {"contact": ["Could not determine type of contact."]})
-        self.assertNumChanges([])
-
-    def test_partial_update_empty(self):
-        response = self.client.patch(reverse('rolecontact-detail', args=[1]), {}, format='json')
-        self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
-        self.assertNumChanges([])
-
-    def test_put_update(self):
-        url = reverse('rolecontact-detail', args=[1])
-        data = {'contact': {'username': 'new_name', 'email': 'new@test.com'},
-                'contact_role': 'pm'}
-        response = self.client.put(url, data, format='json')
-
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(response.data.get('contact').get('username'), 'new_name')
-        self.assertEqual(response.data.get('contact').get('email'), 'new@test.com')
-        self.assertEqual(response.data.get('contact_role'), 'pm')
-        self.assertNumChanges([2])
-
-    def test_put_update_with_bad_data(self):
-        url = reverse('rolecontact-detail', args=[1])
-        data = {'contact': {'user_name': 'new_name', 'email': 'new@test.com'},
-                'contact_role': 'pm'}
-        response = self.client.put(url, data, format='json')
-
-        self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
-        self.assertEqual(response.data, {'contact': ['Could not determine type of contact.']})
-        self.assertNumChanges([])
-
-    def test_delete(self):
-        url = reverse('rolecontact-detail', args=[1])
-        response = self.client.delete(url, format='json')
-
-        self.assertEqual(response.status_code, status.HTTP_204_NO_CONTENT)
-        response = self.client.get(url, format='json')
-        self.assertEqual(response.status_code, status.HTTP_404_NOT_FOUND)
-        self.assertNumChanges([1])
-
-    @mock.patch('rest_framework.mixins.DestroyModelMixin.destroy')
-    def test_delete_protect(self, mock_destory):
-        mock_destory.side_effect = ProtectedError("fake PE", None)
-        url = reverse('rolecontact-detail', args=[1])
-        response = self.client.delete(url, format='json')
-        self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
-        self.assertEqual(response.data, {"detail": "fake PE None"})
         self.assertNumChanges([])
 
 
@@ -889,3 +571,539 @@ class PersonBulkRESTTestCase(TestCaseWithChangeSetMixin, APITestCase):
     def test_partial_update_empty(self):
         response = self.client.patch(reverse('person-list'), {}, format='json')
         self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
+
+
+class GlobalComponentContactRESTTestCase(TestCaseWithChangeSetMixin, APITestCase):
+    fixtures = [
+        "pdc/apps/component/fixtures/tests/global_component.json",
+        "pdc/apps/contact/fixtures/tests/contact_role.json",
+        "pdc/apps/contact/fixtures/tests/maillist.json",
+        "pdc/apps/contact/fixtures/tests/person.json",
+        "pdc/apps/component/fixtures/tests/upstream.json"
+    ]
+
+    @classmethod
+    def setUpTestData(cls):
+        from pdc.apps.component import models as component_models
+        GlobalComponentContact.objects.create(
+            component=component_models.GlobalComponent.objects.get(name='MySQL-python'),
+            role=ContactRole.objects.get(name='pm'),
+            contact=Person.objects.get(username='person1'))
+        GlobalComponentContact.objects.create(
+            component=component_models.GlobalComponent.objects.get(name='java'),
+            role=ContactRole.objects.get(name='qe_ack'),
+            contact=Maillist.objects.get(mail_name="maillist2"))
+
+    def setUp(self):
+        self.list_url = reverse('globalcomponentcontacts-list')
+
+    def test_list_global_component_contacts(self):
+        response = self.client.get(self.list_url)
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        self.assertEqual(response.data.get('count'), 2)
+        results = response.data.get('results')
+        self.assertEqual(results[0]['role'], 'pm')
+        self.assertEqual(results[0]['contact']['email'], 'person1@test.com')
+
+        self.assertEqual(results[1]['role'], 'qe_ack')
+        self.assertEqual(results[1]['contact']['mail_name'], 'maillist2')
+
+    def test_retrieve_global_component_contacts(self):
+        response = self.client.get(reverse('globalcomponentcontacts-detail', args=[2]))
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        self.assertEqual(response.data['role'], 'qe_ack')
+        self.assertEqual(response.data['contact']['mail_name'], 'maillist2')
+        self.assertEqual(response.data['contact']['email'], 'maillist2@test.com')
+
+    def test_destroy_global_component_contacts(self):
+        self.client.delete(reverse('globalcomponentcontacts-detail', args=[2]))
+        self.assertEqual(0, GlobalComponentContact.objects.filter(pk=2).count())
+
+    def test_filter_global_component_contacts(self):
+        response = self.client.get(self.list_url, {'role': 'pm'})
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        self.assertEqual(response.data['count'], 1)
+
+        response = self.client.get(self.list_url, {'role': 'no_such_role'})
+        self.assertEqual(response.data['count'], 0)
+
+        response = self.client.get(self.list_url,
+                                   {'role': 'qe_ack', 'contact': 'maillist2'})
+        self.assertEqual(response.data['count'], 1)
+
+        # should not get anything
+        response = self.client.get(self.list_url,
+                                   {'role': 'qe_ack', 'contact': 'person1'})
+        self.assertEqual(response.data['count'], 0)
+
+        # should not get anything
+        response = self.client.get(self.list_url,
+                                   {'role': 'pm', 'contact': 'maillist2'})
+        self.assertEqual(response.data['count'], 0)
+
+        response = self.client.get(self.list_url, {'component': 'java'})
+        self.assertEqual(response.data['count'], 1)
+
+        response = self.client.get(self.list_url, {'component': 'no_component'})
+        self.assertEqual(response.data['count'], 0)
+
+    def test_create_global_component_contacts(self):
+        data = {'component': 'python', 'role': 'pm', 'contact': {'mail_name': 'maillist2'}}
+        response = self.client.post(self.list_url, data, format='json')
+        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
+        self.assertEqual(3, GlobalComponentContact.objects.count())
+
+    def test_patch_global_component_contacts(self):
+        data = {"contact": {"username": "person2", "email": "person2@test.com"}}
+        response = self.client.patch(reverse('globalcomponentcontacts-detail', args=[2]),
+                                     data, format='json')
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        response.data['contact'].pop('id')
+        self.assertEqual(response.data['contact'], data['contact'])
+        self.assertNumChanges([1])
+
+    def test_create_global_component_contacts_exceed_count_limit(self):
+        data = {'component': 'python', 'role': 'cc', 'contact': {'mail_name': 'maillist2'}}
+        response = self.client.post(self.list_url, data, format='json')
+        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
+
+        data = {'component': 'python', 'role': 'cc', 'contact': {'mail_name': 'maillist1'}}
+        response = self.client.post(self.list_url, data, format='json')
+        self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
+
+    def test_create_global_component_contacts_for_unlimited_role(self):
+        data = {'component': 'python', 'role': 'watcher', 'contact': {'mail_name': 'maillist2'}}
+        response = self.client.post(self.list_url, data, format='json')
+        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
+
+        data = {'component': 'python', 'role': 'watcher', 'contact': {'mail_name': 'maillist1'}}
+        response = self.client.post(self.list_url, data, format='json')
+        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
+
+        data = {'component': 'python', 'role': 'watcher',
+                'contact': {"username": "person1", "email": "person1@test.com"}}
+        response = self.client.post(self.list_url, data, format='json')
+        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
+
+        data = {'component': 'python', 'role': 'watcher',
+                'contact': {"username": "person2", "email": "person2@test.com"}}
+        response = self.client.post(self.list_url, data, format='json')
+        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
+
+    def test_create_global_component_contacts_for_role_limit_3(self):
+        data = {'component': 'python', 'role': 'allow_3_role', 'contact': {'mail_name': 'maillist2'}}
+        response = self.client.post(self.list_url, data, format='json')
+        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
+
+        data = {'component': 'python', 'role': 'allow_3_role', 'contact': {'mail_name': 'maillist1'}}
+        response = self.client.post(self.list_url, data, format='json')
+        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
+
+        data = {'component': 'python', 'role': 'allow_3_role',
+                'contact': {"username": "person1", "email": "person1@test.com"}}
+        response = self.client.post(self.list_url, data, format='json')
+        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
+
+        data = {'component': 'python', 'role': 'allow_3_role',
+                'contact': {"username": "person2", "email": "person2@test.com"}}
+        response = self.client.post(self.list_url, data, format='json')
+        self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
+
+    def test_patch_global_component_contact_exceed_count_limit(self):
+        data = {'component': 'python', 'role': 'cc', 'contact': {'mail_name': 'maillist2'}}
+        response = self.client.post(self.list_url, data, format='json')
+        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
+        data = {'component': 'python', 'role': 'watcher', 'contact': {'mail_name': 'maillist2'}}
+        response = self.client.post(self.list_url, data, format='json')
+        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
+        pk = response.data['id']
+        url = reverse('globalcomponentcontacts-detail', args=[pk])
+        data = {'role': 'cc'}
+        response = self.client.patch(url, data, format='json')
+        self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
+
+    def test_patch_global_component_contacts_for_role_limit_3(self):
+        data = {'component': 'python', 'role': 'allow_3_role', 'contact': {'mail_name': 'maillist2'}}
+        response = self.client.post(self.list_url, data, format='json')
+        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
+
+        data = {'component': 'python', 'role': 'allow_3_role', 'contact': {'mail_name': 'maillist1'}}
+        response = self.client.post(self.list_url, data, format='json')
+        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
+
+        data = {'component': 'python', 'role': 'cc',
+                'contact': {"username": "person1", "email": "person1@test.com"}}
+        response = self.client.post(self.list_url, data, format='json')
+        pk_cc = response.data['id']
+        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
+
+        data = {'component': 'python', 'role': 'watcher',
+                'contact': {"username": "person2", "email": "person2@test.com"}}
+        response = self.client.post(self.list_url, data, format='json')
+        pk_watcher = response.data['id']
+
+        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
+
+        # fourth for allow_3_role
+        url = reverse('globalcomponentcontacts-detail', args=[pk_cc])
+        data = {'role': 'allow_3_role'}
+        response = self.client.patch(url, data, format='json')
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+
+        url = reverse('globalcomponentcontacts-detail', args=[pk_watcher])
+        data = {'role': 'allow_3_role'}
+        response = self.client.patch(url, data, format='json')
+        self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
+
+        # remove one and patch again
+        response = self.client.delete(reverse('globalcomponentcontacts-detail', args=[pk_cc]))
+        self.assertEqual(response.status_code, status.HTTP_204_NO_CONTENT)
+
+        url = reverse('globalcomponentcontacts-detail', args=[pk_watcher])
+        data = {'role': 'allow_3_role'}
+        response = self.client.patch(url, data, format='json')
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+
+    def test_constraint_to_contact_role_count_limit_change(self):
+        data = {'component': 'python', 'role': 'allow_3_role', 'contact': {'mail_name': 'maillist2'}}
+        response = self.client.post(self.list_url, data, format='json')
+        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
+
+        data = {'component': 'python', 'role': 'allow_3_role', 'contact': {'mail_name': 'maillist1'}}
+        response = self.client.post(self.list_url, data, format='json')
+        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
+
+        data = {'component': 'python', 'role': 'allow_3_role',
+                'contact': {"username": "person1", "email": "person1@test.com"}}
+        response = self.client.post(self.list_url, data, format='json')
+        pk = response.data['id']
+        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
+
+        contact_role_url = reverse('contactrole-detail', args=['allow_3_role'])
+        data = {'count_limit': '4'}
+        response = self.client.patch(contact_role_url, data, format='json')
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+
+        contact_role_url = reverse('contactrole-detail', args=['allow_3_role'])
+        data = {'count_limit': '2'}
+        response = self.client.patch(contact_role_url, data, format='json')
+        self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
+
+        # remove one role contact and try again
+        response = self.client.delete(reverse('globalcomponentcontacts-detail', args=[pk]))
+        self.assertEqual(response.status_code, status.HTTP_204_NO_CONTENT)
+        data = {'count_limit': '2'}
+        response = self.client.patch(contact_role_url, data, format='json')
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+
+    def test_constraint_to_contact_role_count_limit_change_for_different_roles(self):
+        data = {'component': 'python', 'role': 'allow_3_role', 'contact': {'mail_name': 'maillist2'}}
+        response = self.client.post(self.list_url, data, format='json')
+        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
+
+        data = {'component': 'python', 'role': 'allow_3_role', 'contact': {'mail_name': 'maillist1'}}
+        response = self.client.post(self.list_url, data, format='json')
+        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
+
+        data = {'component': 'python', 'role': 'allow_3_role',
+                'contact': {"username": "person1", "email": "person1@test.com"}}
+        response = self.client.post(self.list_url, data, format='json')
+        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
+
+        data = {'component': 'python', 'role': 'cc',
+                'contact': {"username": "person1", "email": "person1@test.com"}}
+        response = self.client.post(self.list_url, data, format='json')
+        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
+
+        contact_role_url = reverse('contactrole-detail', args=['allow_3_role'])
+        data = {'count_limit': '2'}
+        response = self.client.patch(contact_role_url, data, format='json')
+        self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
+
+        contact_role_url = reverse('contactrole-detail', args=['cc'])
+        data = {'count_limit': '2'}
+        response = self.client.patch(contact_role_url, data, format='json')
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+
+        contact_role_url = reverse('contactrole-detail', args=['cc'])
+        data = {'count_limit': '1'}
+        response = self.client.patch(contact_role_url, data, format='json')
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+
+
+class ReleaseComponentContactRESTTestCase(TestCaseWithChangeSetMixin, APITestCase):
+    fixtures = [
+        "pdc/apps/component/fixtures/tests/global_component.json",
+        "pdc/apps/release/fixtures/tests/release.json",
+        "pdc/apps/component/fixtures/tests/release_component.json",
+        "pdc/apps/contact/fixtures/tests/contact_role.json",
+        "pdc/apps/contact/fixtures/tests/maillist.json",
+        "pdc/apps/contact/fixtures/tests/person.json",
+        "pdc/apps/component/fixtures/tests/upstream.json"
+    ]
+
+    @classmethod
+    def setUpTestData(cls):
+        from pdc.apps.component import models as component_models
+        ReleaseComponentContact.objects.create(
+            component=component_models.ReleaseComponent.objects.get(name='MySQL-python'),
+            role=ContactRole.objects.get(name='pm'),
+            contact=Person.objects.get(username='person1'))
+        ReleaseComponentContact.objects.create(
+            component=component_models.ReleaseComponent.objects.get(name='python27'),
+            role=ContactRole.objects.get(name='qe_ack'),
+            contact=Maillist.objects.get(mail_name='maillist2'))
+
+    def setUp(self):
+        self.list_url = reverse('releasecomponentcontacts-list')
+
+    def test_list_release_component_contacts(self):
+        response = self.client.get(self.list_url)
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        self.assertEqual(response.data.get('count'), 2)
+        results = response.data.get('results')
+        for result in results:
+            if result['id'] == 1:
+                self.assertEqual(result['role'], 'pm')
+                self.assertEqual(result['contact']['email'], 'person1@test.com')
+            else:
+                self.assertEqual(result['role'], 'qe_ack')
+                self.assertEqual(result['contact']['mail_name'], 'maillist2')
+
+    def test_retrieve_release_component_contacts(self):
+        response = self.client.get(reverse('releasecomponentcontacts-detail', args=[2]))
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        self.assertEqual(response.data['role'], 'qe_ack')
+        self.assertEqual(response.data['contact']['mail_name'], 'maillist2')
+        self.assertEqual(response.data['contact']['email'], 'maillist2@test.com')
+
+    def test_destroy_release_component_contact(self):
+        response = self.client.delete(reverse('releasecomponentcontacts-detail', args=[2]))
+        self.assertEqual(response.status_code, status.HTTP_204_NO_CONTENT)
+        self.assertEqual(0, ReleaseComponentContact.objects.filter(pk=2).count())
+
+    def test_filter_release_component_contacts(self):
+        response = self.client.get(self.list_url, {'role': 'pm'})
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        self.assertEqual(response.data['count'], 1)
+
+        response = self.client.get(self.list_url, {'role': 'no_such_role'})
+        self.assertEqual(response.data['count'], 0)
+
+        response = self.client.get(self.list_url, {'role': 'qe_ack', 'contact': 'maillist2'})
+        self.assertEqual(response.data['count'], 1)
+
+        # should not get anything
+        response = self.client.get(self.list_url, {'role': 'qe_ack', 'contact': 'person1'})
+        self.assertEqual(response.data['count'], 0)
+
+        # should not get anything
+        response = self.client.get(self.list_url, {'role': 'pm', 'contact': 'maillist2'})
+        self.assertEqual(response.data['count'], 0)
+
+        response = self.client.get(self.list_url, {'component': 'python27'})
+        self.assertEqual(response.data['count'], 1)
+
+        response = self.client.get(self.list_url, {'component': 'missing'})
+        self.assertEqual(response.data['count'], 0)
+
+    def test_filter_by_global_component(self):
+        response = self.client.get(self.list_url, {'global_component': 'python'})
+        self.assertEqual(response.data['count'], 1)
+
+    def test_filter_by_email(self):
+        response = self.client.get(self.list_url, {'email': 'person1@test.com'})
+        self.assertEqual(response.data['count'], 1)
+
+        response = self.client.get(self.list_url, {'email': 'maillist2@test.com'})
+        self.assertEqual(response.data['count'], 1)
+
+        response = self.client.get(self.list_url,
+                                   {'email': ['maillist2@test.com', 'person1@test.com']})
+        self.assertEqual(response.data['count'], 2)
+
+        response = self.client.get(self.list_url, {'email': 'bad@test.com'})
+        self.assertEqual(response.data['count'], 0)
+
+    def test_create_release_component_contacts(self):
+        data = {'component': {'id': 2}, 'role': 'cc', 'contact': {'mail_name': 'maillist2'}}
+        response = self.client.post(self.list_url, data, format='json')
+        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
+        self.assertEqual(3, ReleaseComponentContact.objects.count())
+
+    def test_patch_release_component_contacts(self):
+        url = reverse('releasecomponentcontacts-detail', args=[2])
+        data = {"contact": {"username": "person2", "email": "person2@test.com"}}
+        response = self.client.patch(url, data, format='json')
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        response.data['contact'].pop('id')
+        self.assertEqual(response.data['contact'], data['contact'])
+        self.assertNumChanges([1])
+
+    def test_create_release_component_contacts_exceed_count_limit(self):
+        data = {'component': {'id': 2}, 'role': 'cc', 'contact': {'mail_name': 'maillist2'}}
+        response = self.client.post(self.list_url, data, format='json')
+        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
+
+        data = {'component': {'id': 2}, 'role': 'cc', 'contact': {'mail_name': 'maillist1'}}
+        response = self.client.post(self.list_url, data, format='json')
+        self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
+
+    def test_create_release_component_contacts_for_unlimited_role(self):
+        data = {'component': {'id': 2}, 'role': 'watcher', 'contact': {'mail_name': 'maillist2'}}
+        response = self.client.post(self.list_url, data, format='json')
+        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
+
+        data = {'component': {'id': 2}, 'role': 'watcher', 'contact': {'mail_name': 'maillist1'}}
+        response = self.client.post(self.list_url, data, format='json')
+        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
+
+        data = {'component': {'id': 2}, 'role': 'watcher',
+                'contact': {"username": "person1", "email": "person1@test.com"}}
+        response = self.client.post(self.list_url, data, format='json')
+        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
+
+        data = {'component': {'id': 2}, 'role': 'watcher',
+                'contact': {"username": "person2", "email": "person2@test.com"}}
+        response = self.client.post(self.list_url, data, format='json')
+        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
+
+    def test_create_release_component_contacts_for_role_limit_3(self):
+        data = {'component': {'id': 2}, 'role': 'allow_3_role', 'contact': {'mail_name': 'maillist2'}}
+        response = self.client.post(self.list_url, data, format='json')
+        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
+
+        data = {'component': {'id': 2}, 'role': 'allow_3_role', 'contact': {'mail_name': 'maillist1'}}
+        response = self.client.post(self.list_url, data, format='json')
+        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
+
+        data = {'component': {'id': 2}, 'role': 'allow_3_role',
+                'contact': {"username": "person1", "email": "person1@test.com"}}
+        response = self.client.post(self.list_url, data, format='json')
+        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
+
+        data = {'component': {'id': 2}, 'role': 'allow_3_role',
+                'contact': {"username": "person2", "email": "person2@test.com"}}
+        response = self.client.post(self.list_url, data, format='json')
+        self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
+
+    def test_patch_release_component_contact_exceed_count_limit(self):
+        data = {'component': {'id': 2}, 'role': 'cc', 'contact': {'mail_name': 'maillist2'}}
+        response = self.client.post(self.list_url, data, format='json')
+        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
+        data = {'component': {'id': 2}, 'role': 'watcher', 'contact': {'mail_name': 'maillist2'}}
+        response = self.client.post(self.list_url, data, format='json')
+        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
+        pk = response.data['id']
+        url = reverse('releasecomponentcontacts-detail', args=[pk])
+        data = {'role': 'cc'}
+        response = self.client.patch(url, data, format='json')
+        self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
+
+    def test_patch_release_component_contacts_for_role_limit_3(self):
+        data = {'component': {'id': 2}, 'role': 'allow_3_role', 'contact': {'mail_name': 'maillist2'}}
+        response = self.client.post(self.list_url, data, format='json')
+        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
+
+        data = {'component': {'id': 2}, 'role': 'allow_3_role', 'contact': {'mail_name': 'maillist1'}}
+        response = self.client.post(self.list_url, data, format='json')
+        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
+
+        data = {'component': {'id': 2}, 'role': 'cc',
+                'contact': {"username": "person1", "email": "person1@test.com"}}
+        response = self.client.post(self.list_url, data, format='json')
+        pk_cc = response.data['id']
+        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
+
+        data = {'component': {'id': 2}, 'role': 'watcher',
+                'contact': {"username": "person2", "email": "person2@test.com"}}
+        response = self.client.post(self.list_url, data, format='json')
+        pk_watcher = response.data['id']
+
+        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
+
+        # fourth for allow_3_role
+        url = reverse('releasecomponentcontacts-detail', args=[pk_cc])
+        data = {'role': 'allow_3_role'}
+        response = self.client.patch(url, data, format='json')
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+
+        url = reverse('releasecomponentcontacts-detail', args=[pk_watcher])
+        data = {'role': 'allow_3_role'}
+        response = self.client.patch(url, data, format='json')
+        self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
+
+        # remove one and patch again
+        response = self.client.delete(reverse('releasecomponentcontacts-detail', args=[pk_cc]))
+        self.assertEqual(response.status_code, status.HTTP_204_NO_CONTENT)
+
+        url = reverse('releasecomponentcontacts-detail', args=[pk_watcher])
+        data = {'role': 'allow_3_role'}
+        response = self.client.patch(url, data, format='json')
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+
+    def test_constraint_to_contact_role_count_limit_change(self):
+        data = {'component': {'id': 2}, 'role': 'allow_3_role', 'contact': {'mail_name': 'maillist2'}}
+        response = self.client.post(self.list_url, data, format='json')
+        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
+
+        data = {'component': {'id': 2}, 'role': 'allow_3_role', 'contact': {'mail_name': 'maillist1'}}
+        response = self.client.post(self.list_url, data, format='json')
+        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
+
+        data = {'component': {'id': 2}, 'role': 'allow_3_role',
+                'contact': {"username": "person1", "email": "person1@test.com"}}
+        response = self.client.post(self.list_url, data, format='json')
+        pk = response.data['id']
+        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
+
+        contact_role_url = reverse('contactrole-detail', args=['allow_3_role'])
+        data = {'count_limit': '4'}
+        response = self.client.patch(contact_role_url, data, format='json')
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+
+        contact_role_url = reverse('contactrole-detail', args=['allow_3_role'])
+        data = {'count_limit': '2'}
+        response = self.client.patch(contact_role_url, data, format='json')
+        self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
+
+        # remove one role contact and try again
+        response = self.client.delete(reverse('releasecomponentcontacts-detail', args=[pk]))
+        self.assertEqual(response.status_code, status.HTTP_204_NO_CONTENT)
+        data = {'count_limit': '2'}
+        response = self.client.patch(contact_role_url, data, format='json')
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+
+    def test_constraint_to_contact_role_count_limit_change_for_different_roles(self):
+        data = {'component': {'id': 2}, 'role': 'allow_3_role', 'contact': {'mail_name': 'maillist2'}}
+        response = self.client.post(self.list_url, data, format='json')
+        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
+
+        data = {'component': {'id': 2}, 'role': 'allow_3_role', 'contact': {'mail_name': 'maillist1'}}
+        response = self.client.post(self.list_url, data, format='json')
+        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
+
+        data = {'component': {'id': 2}, 'role': 'allow_3_role',
+                'contact': {"username": "person1", "email": "person1@test.com"}}
+        response = self.client.post(self.list_url, data, format='json')
+        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
+
+        data = {'component': {'id': 2}, 'role': 'cc',
+                'contact': {"username": "person1", "email": "person1@test.com"}}
+        response = self.client.post(self.list_url, data, format='json')
+        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
+
+        contact_role_url = reverse('contactrole-detail', args=['allow_3_role'])
+        data = {'count_limit': '2'}
+        response = self.client.patch(contact_role_url, data, format='json')
+        self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
+
+        contact_role_url = reverse('contactrole-detail', args=['cc'])
+        data = {'count_limit': '2'}
+        response = self.client.patch(contact_role_url, data, format='json')
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+
+        contact_role_url = reverse('contactrole-detail', args=['cc'])
+        data = {'count_limit': '1'}
+        response = self.client.patch(contact_role_url, data, format='json')
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
